dspy-pattern/langgraph_python_example/requirements.txt
dspy-ai
requests
python-dotenv
